en withDraw y payDividend cambie el tipo del parametro pay de int a double

+++ src/ufps/poo2/ejercicio/banco/modelo/Bank.java
@@ -38,11 +38,11 @@ public class Bank {
 
     }
 
-    public void withdrawAccount(int accnum, int pay) {
+    public void withdrawAccount(int accnum, double pay) {
         buscarCuenta(accnum).withdraw(pay);
     }
 
-    public void payDividend(int accnum, int pay) {
+    public void payDividend(int accnum, double pay) {
         buscarCuenta(accnum).deposit(pay);
 
     }
